fix: реализовал использование фун-ии getUserIDFromToken из baseHandler'а
