+++ handlers/createChat.go
@@ -4,27 +4,11 @@ import (
 	"atnidirtysleng/db"
 	"net/http"
 
-	"fmt"
 	"log"
 
-	"github.com/dgrijalva/jwt-go"
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
 )
-
-func getOwnerFromToken(tokenString string) (string, error) {
-	token, err := parseToken(tokenString)
-	if err != nil {
-		return "", err
-	}
-
-	if claims, ok := token.Claims.(jwt.MapClaims); ok && token.Valid {
-		owner := claims["id"].(string)
-		return owner, nil
-	}
-
-	return "", fmt.Errorf("Invalid token")
-}
 
 func (h BaseHandler) CreateChat(c *gin.Context) {
 	var chat db.Chat
@@ -40,7 +24,7 @@ func (h BaseHandler) CreateChat(c *gin.Context) {
 	}
 
 	tokenString := c.GetHeader("Authorization")
-	ownerStr, err := getOwnerFromToken(tokenString)
+	ownerStr, err := getUserIDFromToken(tokenString)
 	if err != nil {
 		log.Println("Error extracting owner from token:", err.Error())
 		c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
